Move utils out of components/dashboard.
