+++ assets/js/components/publisher-wins/total-stats.js
@@ -24,7 +24,7 @@ import { __ } from '@wordpress/i18n';
 /**
  * Internal dependencies
  */
-import { extractSearchConsoleDashboardData } from '../../modules/search-console/components/dashboard/util';
+import { extractSearchConsoleDashboardData } from '../../modules/search-console/dashboard/util';
 import { getTimeInSeconds, getDaysBetweenDates } from '../../util';
 
 const totalStats = ( data, id ) => {
