+++ assets/js/modules/search-console/components/dashboard/DashboardWidgetPopularKeywordsTable.js
@@ -38,7 +38,7 @@ import PreviewTable from '../../../../components/preview-table';
 import Layout from '../../../../components/layout/layout';
 import {
 	isDataZeroSearchConsole,
-} from './util';
+} from '../../dashboard/util';
 
 class DashboardWidgetPopularKeywordsTable extends Component {
 	static renderLayout( component ) {
