+++ assets/js/modules/search-console/components/dashboard/SearchConsoleDashboardWidgetSiteStats.js
@@ -30,7 +30,7 @@ import withData from '../../../../components/higherorder/withdata';
 import { TYPE_MODULES } from '../../../../components/data';
 import GoogleChart from '../../../../components/google-chart.js';
 import PreviewBlock from '../../../../components/preview-block';
-import { extractSearchConsoleDashboardData } from './util';
+import { extractSearchConsoleDashboardData } from '../../dashboard/util';
 
 class SearchConsoleDashboardWidgetSiteStats extends Component {
 	constructor( props ) {
