+++ assets/js/modules/search-console/components/wp-dashboard/WPSearchConsoleDashboardWidgetOverview.js
@@ -33,7 +33,7 @@ import { TYPE_MODULES } from '../../../../components/data';
 import {
 	extractSearchConsoleDashboardData,
 	isDataZeroSearchConsole,
-} from '../dashboard/util';
+} from '../../dashboard/utiltil';
 import CTA from '../../../../components/notifications/cta';
 
 class WPSearchConsoleDashboardWidgetOverview extends Component {
